# Improve the worst case performance of Cursor::next()