As mentioned in a TODO, try to walk up the cached subtree and skip the 0-measure
nodes. I checked that the two implementations coincide by walking line by line some
files.

It degrades a little the performance when the lines are short, for example in
the case of the sqlite source:
```
after    ... bench:   4,530,490 ns/iter (+/- 733,470)
before   ... bench:   4,128,834 ns/iter (+/- 411,246)
```
However, for a ~30MB file with a single newline at the end, it is significantly faster:
```
after    ... bench:       9,794 ns/iter (+/- 596)
before   ... bench:   2,388,631 ns/iter (+/- 318,327)
```