@ineol 2017-05-10
Thank you for the comments.

It appears that Cursor::prev is doesn't move when it is already is on a boundary, so I don't see an obvious way to test it; this is why I only wrote tests/benchmarks for next.

Now that I think of it, with Cursor::set I can write tests but no benchmarks...