@ineol 2017-05-10
For information, the benchmark situation is as follows on my computer (the numbers seem stable accros runs):

before:
```
test benchmark_few_big_lines ... bench:     753,114 ns/iter (+/- 36,710)
test benchmark_short_lines   ... bench:  16,673,272 ns/iter (+/- 863,349)
test benchmark_triangle      ... bench: 110,738,718 ns/iter (+/- 15,690,231)
```
without inlining:
```
test benchmark_few_big_lines ... bench:       4,365 ns/iter (+/- 1,869)
test benchmark_short_lines   ... bench:  22,283,243 ns/iter (+/- 1,538,873)
test benchmark_triangle      ... bench:  22,053,478 ns/iter (+/- 1,272,111)
```
with inlining:
```
test benchmark_few_big_lines ... bench:       4,288 ns/iter (+/- 1,886)
test benchmark_short_lines   ... bench:  18,841,222 ns/iter (+/- 949,549)
test benchmark_triangle      ... bench:  22,529,569 ns/iter (+/- 1,679,518)
```