@ineol 2017-05-10
About `prev`, it seems to stop just *after* the boundary, so that the next time it is called it does not need to move.