@raphlinus 2017-05-13
Ok, looks good enough now to merge. There's definitely follow-up work on `prev` and possibly some more doc cleanup, but this is good forward progress. Thanks!